Move mssql queries
