+++ ingestion/src/metadata/utils/sql_queries.py
@@ -14,32 +14,6 @@ SQL Queries used during ingestion
 
 import textwrap
 
-MSSQL_SQL_STATEMENT = textwrap.dedent(
-    """
-      SELECT TOP {result_limit}
-        db.NAME database_name,
-        t.text query_text,
-        s.last_execution_time start_time,
-        DATEADD(s, s.total_elapsed_time/1000, s.last_execution_time) end_time,
-        NULL schema_name,
-        NULL query_type,
-        NULL user_name,
-        NULL aborted
-      FROM sys.dm_exec_cached_plans AS p
-      INNER JOIN sys.dm_exec_query_stats AS s
-        ON p.plan_handle = s.plan_handle
-      CROSS APPLY sys.Dm_exec_sql_text(p.plan_handle) AS t
-      INNER JOIN sys.databases db
-        ON db.database_id = t.dbid
-      WHERE s.last_execution_time between '{start_time}' and '{end_time}'
-          AND t.text NOT LIKE '/* {{"app": "OpenMetadata", %%}} */%%'
-          AND t.text NOT LIKE '/* {{"app": "dbt", %%}} */%%'
-          AND p.objtype != 'Prepared'
-          {filters}
-      ORDER BY s.last_execution_time DESC
-"""
-)
-
 CLICKHOUSE_SQL_STATEMENT = textwrap.dedent(
     """
         Select
